halfway changes, added ssr, need to clear slug, and make sure playerid in rigiht place
